+++ R/DataSpaceConnection.R
@@ -133,7 +133,7 @@ DataSpaceConnection <- R6Class(
     getStudies = function(availableStudies = self$availableStudies) {
       if(is.character(availableStudies))
         availableStudies <- self$availableStudies[study_id %in% availableStudies]
-      if(!"availableStudies" %in% class(availableStudies))
+      if(!"availableStudies" %in% attr(availableStudies, "dsr_type"))
         stop("Argument must be `character` or `availableStudies`.")
       if(nrow(availableStudies) == 0)
         stop("No vaild `availableStudies`, or `study_id` vector passed.")
@@ -147,7 +147,7 @@ DataSpaceConnection <- R6Class(
     getGroups = function(availableGroups = self$availableGroups) {
       if(is.numeric(availableGroups))
         availableGroups <- self$availableGroups[group_id %in% availableGroups]
-      if(!"availableGroups" %in% class(availableGroups))
+      if(!"availableGroups" %in% attr(availableGroups, "dsr_type"))
         stop("Argument must be `numeric` or `avilableGroups`.")
       if(nrow(availableGroups) == 0)
         stop("No vaild `availableGroups`, or `group_id` vector passed.")
@@ -171,7 +171,7 @@ DataSpaceConnection <- R6Class(
           )
         ]
 
-      if(!any(class(availableMabs) %in% c("availableMabs", "availableMabMixtures")))
+      if(!any(attr(availableMabs, "dsr_type") %in% c("availableMabs", "availableMabMixtures")))
         stop("Argument must be `character`, `availableMabs`, or `availableMabMixtures`.")
 
       if(length(includeMixtures) != 1 || (!includeMixtures %in% c("yes", "no", "only")))
@@ -180,7 +180,7 @@ DataSpaceConnection <- R6Class(
       if(nrow(availableMabs) == 0)
         stop("No mAbs available from `availableMabs` argument.")
 
-      if("availableMabMixtures" %in% class(availableMabs)){
+      if("availableMabMixtures" %in% attr(availableMabs, "dsr_type")){
         message("All mabs in all mixtures passed with be returned.")
         availableMabs <-
           merge(availableMabs, private$.shared$.mabMix, by= "mab_mix_id")
@@ -196,7 +196,7 @@ DataSpaceConnection <- R6Class(
       if(is.character(availableDonors))
         availableGroups <- self$availableDonors[donor_id %in% availableDonors]
 
-      if(!"availableDonors" %in% class(self$availableDonors))
+      if(!"availableDonors" %in% attr(self$availableDonors, "dsr_type"))
         stop("Argument must be `character` or `availableDonors`.")
 
       if(nrow(availableDonors) == 0)
@@ -222,7 +222,7 @@ DataSpaceConnection <- R6Class(
       if(is.character(availablePublications))
         availablePublications <- self$availablePublications[publication_id == availablePublications]
 
-      if(!"availablePublications" %in% class(availablePublications))
+      if(!"availablePublications" %in% attr(availablePublications, "dsr_type"))
         stop("Argument must be `character` or `availablePublications`.")
 
       if(nrow(availablePublications) == 0)
@@ -242,6 +242,43 @@ DataSpaceConnection <- R6Class(
         warning("Not all publication ids successfully downloaded: ", paste(downloadDocument[success == FALSE, publication_id]))
 
       return(invisible(downloadDocument$destination))
+
+    },
+
+    #' @description
+    #' Load any available mAb metadata from LANL.
+    loadLanlMabMetadata = function(){
+
+      if(is.null(private$.mabIds)){
+        stop("This objects does not load LANL mAb metadata.")
+      } else {
+        if(!is.null(private$.shared$.lanlMabMetadata)) {
+
+          mabIds <- setdiff(
+            private$.shared$.mabMetadata[!is.na(mab_lanl_id) & mab_id %in% private$.mabIds, mab_id],
+            private$.shared$.lanlMabMetadata$mab_id
+          )
+
+        } else {
+          mabIds <- private$.mabIds
+        }
+      }
+
+      mabIds <- private$.shared$.mabMetadata[!is.na(mab_lanl_id) & mab_id %in% mabIds, mab_id]
+
+      if(length(mabIds) > 0){
+        private$.shared$.lanlMabMetadata <- rbind(
+          private$.shared$.lanlMabMetadata,
+          merge(
+            private$.shared$.mabMetadata[,.(mab_id, mab_lanl_id)],
+            fetchProcessLanlMetadata(private$.shared$.mabMetadata[mab_id %in% mabIds]),
+            by.x = "mab_lanl_id",
+            by.y = "id"
+          ),
+          fill = TRUE
+        )
+        setcolorder(private$.shared$.lanlMabMetadata, c("mab_id", "mab_lanl_id"))
+      }
 
     },
 
@@ -332,6 +369,22 @@ DataSpaceConnection <- R6Class(
         publication_id %in% private$.publicationIds,
         -c("url", "remote_path", "document_id")
       ]
+    },
+
+    #' @field lanlMabMetadata A data.table of mAb metadata from LANL
+    #' of mAbs found in the object
+    lanlMabMetadata = function() {
+      if(is.null(private$.shared$.lanlMabMetadata))
+        stop("No LANL mAb metadata loaded. Run the `loadLanlMetadata()` method for this object to load LANL mAb metadata.")
+
+      lanlMabs <- private$.shared$.mabMetadata[!is.na(mab_lanl_id), mab_id]
+
+      if(
+        !all(private$.mabIds[private$.mabIds %in% lanlMabs] %in% private$.shared$.lanlMabMetadata[mab_id %in% lanlMabs,mab_id])
+      )
+        warning("Not all object LANL mabs have metadata loaded from LANL. Run the `loadLanlMetadata()` method for this object to load LANL mAb metadata.")
+
+      private$.shared$.lanlMabMetadata[mab_id %in% private$.mabIds]
     },
 
     #' @field virusNameMappingTables A list of data.tables containing virus name mappings.
@@ -439,7 +492,7 @@ DataSpaceConnection <- R6Class(
         merge(niData, all.x = TRUE)
 
       private$.shared$.availableStudies <- study
-      class(private$.shared$.availableStudies) <- c(class(private$.shared$.availableStudies), "availableStudies")
+      attr(private$.shared$.availableStudies, "dsr_type") <- "availableStudies"
 
     },
     .loadAvailableGroups = function() {
@@ -479,7 +532,7 @@ DataSpaceConnection <- R6Class(
         setorder(group_id) |>
         setkey(group_id)
 
-      class(private$.shared$.availableGroups) <- c(class(private$.shared$.availableGroups), "availableGroups")
+      attr(private$.shared$.availableGroups, "dsr_type") <- "availableGroups"
 
     },
     .loadAvailableDonors = function() {
@@ -512,7 +565,7 @@ DataSpaceConnection <- R6Class(
         donors[,c(names(private$.shared$.donorMetadata), "lineage_sequences_available", grep("_count", names(donors), value = TRUE)), with = FALSE] |>
         unique()
 
-      class(private$.shared$.availableDonors) <- c(class(private$.shared$.availableDonors), "availableDonors")
+      attr(private$.shared$.availableDonors, "dsr_type") <- "availableDonors"
 
     },
     .loadAvailableMabs = function() {
@@ -547,7 +600,7 @@ DataSpaceConnection <- R6Class(
         ), mab_id] |>
         unique()
 
-      class(private$.shared$.availableMabs) <- c(class(private$.shared$.availableMabs), "availableMabs")
+      attr(private$.shared$.availableMabs, "dsr_type") <- "availableMabs"
     
     },
     .loadAvailableMabMixtures = function() {
@@ -577,7 +630,7 @@ DataSpaceConnection <- R6Class(
         ) |>
           unique()
 
-      class(private$.shared$.availableMabMixtures) <- c(class(private$.shared$.availableMabMixtures), "availableMabMixtures")
+      attr(private$.shared$.availableMabMixtures, "dsr_type") <- "availableMabMixtures"
 
     },
     .loadAvailableViruses = function() {
@@ -682,7 +735,7 @@ DataSpaceConnection <- R6Class(
         "&publicAccess=true"
       )]
 
-      class(pubs) <- c(class(pubs), "availablePublications")
+      attr(pubs, "dsr_type") <- "availablePublications"
       private$.shared$.availablePublications <- pubs
 
     },