+++ R/DataSpaceStudies.R
@@ -168,20 +168,28 @@ DataSpaceStudies <- R6Class(
       private$.datasets <- assayData
 
       if("BCR_Sequence" %in% assays){
-        bcr_sequence <- labkey.selectRows(
-          baseUrl = private$.shared$.config$labkeyUrlBase,
-          folderPath="CAVD",
-          schemaName="CDS",
-          queryName="bcr_sequence_alignment_prot",
-          colFilter=makeFilter(c("prot", "IN", paste(private$.studyIds, collapse = ";"))),
-          colNameOpt="rname"
-        ) |> setDT()
+
+        message("Querying BCR sequence study data, this may take a few minutes...")
+
+        bcr_sequence <-
+          Map(
+            \(studyId)
+            labkey.selectRows(
+              baseUrl = private$.shared$.config$labkeyUrlBase,
+              folderPath="CAVD",
+              schemaName="CDS",
+              queryName="bcr_sequence_alignment_prot",
+              colFilter=makeFilter(c("prot", "IN", studyId)),
+              colNameOpt="rname"
+            ) |> setDT(),
+            private$.studyIds
+          ) |> rbindlist()
 
         private$.datasets <- append(
           private$.datasets,
           list(
-            alignments = bcr_sequence[,.(alignment_id,sequence_id,stop_codon,vj_in_frame,productive,rev_comp,complete_vdj,v_call,d_call,j_call,sequence_alignment,germline_alignment,sequence_alignment_aa,germline_alignment_aa,v_alignment_start,v_alignment_end,d_alignment_start,d_alignment_end,j_alignment_start,j_alignment_end,v_sequence_alignment,v_sequence_alignment_aa,v_germline_alignment,v_germline_alignment_aa,d_sequence_alignment,d_sequence_alignment_aa,d_germline_alignment,d_germline_alignment_aa,j_sequence_alignment,j_sequence_alignment_aa,j_germline_alignment,j_germline_alignment_aa,fwr1,fwr1_aa,cdr1,cdr1_aa,fwr2,fwr2_aa,cdr2,cdr2_aa,fwr3,fwr3_aa,fwr4,fwr4_aa,cdr3,cdr3_aa,junction,junction_length,junction_aa,junction_aa_length,v_score,d_score,j_score,v_cigar,d_cigar,j_cigar,v_support,d_support,j_support,v_identity,d_identity,j_identity,v_sequence_start,v_sequence_end,v_germline_start,v_germline_end,d_sequence_start,d_sequence_end,d_germline_start,d_germline_end,j_sequence_start,j_sequence_end,j_germline_start,j_germline_end,fwr1_start,fwr1_end,cdr1_start,cdr1_end,fwr2_start,fwr2_end,cdr2_start,cdr2_end,fwr3_start,fwr3_end,fwr4_start,fwr4_end,cdr3_start,cdr3_end,np1,np1_length,np2,np2_length,run_application,d_frame,c_call,c_alignment_start,c_alignment_end,c_sequence_alignment,c_sequence_alignment_aa,c_germline_alignment,c_germline_alignment_aa,c_score,c_cigar,c_support,c_identity,c_sequence_start,c_sequence_end,c_germline_start,c_germline_end)],
-            sequences = bcr_sequence[,.(sequence_id,sequence_nt,sequence_aa,chain)],
+            alignments = bcr_sequence[,.(cell_id,sequence_id,stop_codon,vj_in_frame,productive,rev_comp,complete_vdj,v_call,d_call,j_call,sequence_alignment,germline_alignment,sequence_alignment_aa,germline_alignment_aa,v_alignment_start,v_alignment_end,d_alignment_start,d_alignment_end,j_alignment_start,j_alignment_end,v_sequence_alignment,v_sequence_alignment_aa,v_germline_alignment,v_germline_alignment_aa,d_sequence_alignment,d_sequence_alignment_aa,d_germline_alignment,d_germline_alignment_aa,j_sequence_alignment,j_sequence_alignment_aa,j_germline_alignment,j_germline_alignment_aa,fwr1,fwr1_aa,cdr1,cdr1_aa,fwr2,fwr2_aa,cdr2,cdr2_aa,fwr3,fwr3_aa,fwr4,fwr4_aa,cdr3,cdr3_aa,junction,junction_length,junction_aa,junction_aa_length,v_score,d_score,j_score,v_cigar,d_cigar,j_cigar,v_support,d_support,j_support,v_identity,d_identity,j_identity,v_sequence_start,v_sequence_end,v_germline_start,v_germline_end,d_sequence_start,d_sequence_end,d_germline_start,d_germline_end,j_sequence_start,j_sequence_end,j_germline_start,j_germline_end,fwr1_start,fwr1_end,cdr1_start,cdr1_end,fwr2_start,fwr2_end,cdr2_start,cdr2_end,fwr3_start,fwr3_end,fwr4_start,fwr4_end,cdr3_start,cdr3_end,np1,np1_length,np2,np2_length,run_application,d_frame,c_call,c_alignment_start,c_alignment_end,c_sequence_alignment,c_sequence_alignment_aa,c_germline_alignment,c_germline_alignment_aa,c_score,c_cigar,c_support,c_identity,c_sequence_start,c_sequence_end,c_germline_start,c_germline_end)],
+            sequences = bcr_sequence[,.(cell_id,sequence_id,sequence_nt,sequence_aa,chain)],
             runInformation = labkey.selectRows(
               baseUrl = private$.shared$.config$labkeyUrlBase,
               folderPath="CAVD",
@@ -246,7 +254,7 @@ DataSpaceStudies <- R6Class(
         private$.availableIntegratedDatasets,
         private$.availableNIDatasets
       )
-      class(datasets) <- c("availableDatasets", class(datasets))
+      attr(datasets, "dsr_type") <- "availableDatasets"
       return(datasets)
     },